Create task.py
