+++ bot/tasksdbs.py
@@ -1,14 +1,5 @@
 import sqlite3
-
-
-class Task(object):
-
-    def __init__(self, name: str, url: str):
-        self.name = name
-        self.url = url
-
-    def show(self):
-        return f"*Task {self.name}*\n {self.url}"
+from task import Task
 
 
 class TasksDatabase(object):
